requirements.txt
azure-ai-formrecognizer==3.0.0
